cw1200: Adds 4byte alignment for 16bit read/writes
This is moreless workaround for mmci 4byte
alignment restrictions.

ST-Ericsson ID: 373107
ST-Ericsson FOSS-OUT ID: NA

Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/39076

+++ drivers/staging/cw1200/bh.c
@@ -265,7 +265,7 @@ static int cw1200_bh(void *arg)
 	int tx_burst;
 	int rx_burst = 0;
 	long status;
-	u8 dummy;
+	u32 dummy;
 
 	for (;;) {
 		if (!priv->hw_bufs_used
